cambios varios, implementación de nuevo controller para producto y correcciones de customer

+++ businessdomain/product/src/main/java/com/paymentchain/product/controller/ProductRestController.java
@@ -1,6 +1,7 @@
 package com.paymentchain.product.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,6 +10,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -28,29 +30,17 @@ public class ProductRestController {
         return productRepository.findAll();
     }
     
-//    @GetMapping("/{id}")
-//    public ResponseEntity<?> get(@PathVariable long id) {
-//         Optional<Product> customer = productRepository.findById(id);
-//        if (customer.isPresent()) {
-//            return new ResponseEntity<>(customer.get(), HttpStatus.OK);
-//        } else {
-//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//        }
-//    }
+    @GetMapping("/{id}")
+    public ResponseEntity<?> get(@PathVariable long id) {
+        Product product = productRepository.findById(id).get();
+        return ResponseEntity.ok(product);
+    }
     
-//    @PutMapping("/{id}")
-//    public ResponseEntity<?> put(@PathVariable long id, @RequestBody Product input) {
-//         Optional<Product> optionalcustomer = productRepository.findById(id);
-//        if (optionalcustomer.isPresent()) {
-//            Product newcustomer = optionalcustomer.get();
-//            newcustomer.setName(input.getName());
-//            newcustomer.setPhone(input.getPhone());
-//             Product save = productRepository.save(newcustomer);
-//          return new ResponseEntity<>(save, HttpStatus.OK);
-//        } else {
-//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//        }
-//    }
+    @PutMapping("/{id}")
+    public ResponseEntity<?> put(@PathVariable long id, @RequestBody Product input) {
+        Product save = productRepository.save(input);
+          return new ResponseEntity<>(save, HttpStatus.OK);
+    }
     
     @PostMapping
     public ResponseEntity<?> post(@RequestBody Product input) {
@@ -60,7 +50,10 @@ public class ProductRestController {
     
     @DeleteMapping("/{id}")
     public ResponseEntity<?> delete(@PathVariable long id) {
-         productRepository.deleteById(id);
+         Optional<Product> product = productRepository.findById(id);
+         if (product != null) {
+        	 productRepository.deleteById(id);
+         }
          return new ResponseEntity<>(HttpStatus.OK);
     }
